alias-help() is now dynamic
instead of maintaining a static help function for alias groups, keep it
DRY by pulling those aliases out as 'metadata' with metafor(). This
requires only a little massaging by adding a trailing quote, which
metafor() strips.

This will now display help for only those aliases which are actually
being used, or optionally for a particular group of aliases.

Examples:
1. $ alias-help         # displays help for all installed aliases
2. $ alias-help git     # displays help for git aliases